Move JMSConnector builder to inner class
